opgops-1012: EBS snapshotting script
A script that can:
  * snapshot all attached EBS volumes to an instance
  * prune N old snapshots
  * runs in either "single or discover" mode

This will allow a single, or all instances in a region to have a
snapshot of all their volumes taken.

Change the default retention period from 7 days by setting the
EBS_RETENTION_PERIOD environment variable.

Note this script requires the awscli, and IAM credentials with the
following policy statement:

{
    "Version": "2012-10-17",
    "Statement": [
        {
            "Sid": "Stmt1426256275000",
            "Effect": "Allow",
            "Action": [
                "ec2:DescribeInstances",
                "ec2:CreateSnapshot",
                "ec2:CreateTags",
                "ec2:DeleteSnapshot",
                "ec2:DescribeSnapshots",
                "ec2:DescribeVolumes"
            ],
            "Resource": [
                "*"
            ]
        }
    ]
}